+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/service/MullvadTileService.kt
@@ -5,14 +5,17 @@ import android.graphics.drawable.Icon
 import android.os.Build
 import android.service.quicksettings.Tile
 import android.service.quicksettings.TileService
-import kotlin.properties.Delegates.observable
-import kotlinx.coroutines.CoroutineScope
+import android.util.Log
 import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.MainScope
-import kotlinx.coroutines.cancel
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.debounce
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
+import kotlinx.coroutines.withTimeoutOrNull
 import net.mullvad.mullvadvpn.R
 import net.mullvad.mullvadvpn.ipc.ServiceConnection
 import net.mullvad.mullvadvpn.model.ServiceResult
@@ -20,47 +23,66 @@ import net.mullvad.mullvadvpn.model.TunnelState
 import net.mullvad.talpid.tunnel.ActionAfterDisconnect
 
 class MullvadTileService : TileService() {
-    private var secured by observable(false) { _, _, _ ->
-        updateTileState()
-    }
-
-    private lateinit var scope: CoroutineScope
+    private val scope = MainScope()
+    private var listenerJob: Job? = null
 
     private lateinit var securedIcon: Icon
     private lateinit var unsecuredIcon: Icon
 
     override fun onCreate() {
-        super.onCreate()
-
         securedIcon = Icon.createWithResource(this, R.drawable.small_logo_white)
         unsecuredIcon = Icon.createWithResource(this, R.drawable.small_logo_black)
     }
 
-    override fun onStartListening() {
-        super.onStartListening()
-
-        scope = MainScope()
-
-        scope.launch { listenToTunnelState() }
-    }
-
     override fun onClick() {
-        super.onClick()
+        // Workaround for the reported bug: https://issuetracker.google.com/issues/236862865
+        suspend fun isUnlockStatusPropagatedWithinTimeout(
+            unlockTimeoutMillis: Long,
+            unlockCheckDelayMillis: Long
+        ): Boolean {
+            return withTimeoutOrNull(unlockTimeoutMillis) {
+                while (isLocked) {
+                    delay(unlockCheckDelayMillis)
+                }
+                return@withTimeoutOrNull true
+            } ?: false
+        }
 
-        val intent = Intent(this, MullvadVpnService::class.java)
+        unlockAndRun {
+            runBlocking {
+                val isUnlockStatusPropagated = isUnlockStatusPropagatedWithinTimeout(
+                    unlockTimeoutMillis = 1000L,
+                    unlockCheckDelayMillis = 100L
+                )
 
-        if (secured) {
-            intent.action = MullvadVpnService.KEY_DISCONNECT_ACTION
-            startService(intent)
-        } else {
-            intent.action = MullvadVpnService.KEY_CONNECT_ACTION
-            startForegroundService(intent)
+                if (isUnlockStatusPropagated) {
+                    toggleTunnel()
+                } else {
+                    Log.e("mullvad", "Unable to toggle tunnel state")
+                }
+            }
         }
     }
 
+    override fun onStartListening() {
+        listenerJob = scope.launch { listenToTunnelState() }
+    }
+
     override fun onStopListening() {
-        scope.cancel()
-        super.onStopListening()
+        listenerJob?.cancel()
+    }
+
+    private fun toggleTunnel() {
+        val intent = Intent(this, MullvadVpnService::class.java).apply {
+            action = if (qsTile.state == Tile.STATE_INACTIVE) {
+                MullvadVpnService.KEY_CONNECT_ACTION
+            } else {
+                MullvadVpnService.KEY_DISCONNECT_ACTION
+            }
+        }
+
+        // Always start as foreground in case tile is out-of-sync.
+        startForegroundService(intent)
     }
 
     @OptIn(FlowPreview::class)
@@ -68,31 +90,44 @@ class MullvadTileService : TileService() {
         ServiceConnection(this@MullvadTileService, scope)
             .tunnelState
             .debounce(300L)
-            .collect { updateTunnelState(it.first, it.second) }
+            .map { (tunnelState, connectionState) -> mapToTileState(tunnelState, connectionState) }
+            .collect {
+                updateTileState(it)
+            }
     }
 
-    private fun updateTunnelState(
+    private fun mapToTileState(
         tunnelState: TunnelState,
         connectionState: ServiceResult.ConnectionState
-    ) {
-        secured = if (connectionState == ServiceResult.ConnectionState.CONNECTED) {
+    ): Int {
+        return if (connectionState == ServiceResult.ConnectionState.CONNECTED) {
             when (tunnelState) {
-                is TunnelState.Disconnected -> false
-                is TunnelState.Connecting -> true
-                is TunnelState.Connected -> true
+                is TunnelState.Disconnected -> Tile.STATE_INACTIVE
+                is TunnelState.Connecting -> Tile.STATE_ACTIVE
+                is TunnelState.Connected -> Tile.STATE_ACTIVE
                 is TunnelState.Disconnecting -> {
-                    tunnelState.actionAfterDisconnect == ActionAfterDisconnect.Reconnect
+                    if (tunnelState.actionAfterDisconnect == ActionAfterDisconnect.Reconnect) {
+                        Tile.STATE_ACTIVE
+                    } else {
+                        Tile.STATE_INACTIVE
+                    }
                 }
-                is TunnelState.Error -> tunnelState.errorState.isBlocking
+                is TunnelState.Error -> {
+                    if (tunnelState.errorState.isBlocking) {
+                        Tile.STATE_ACTIVE
+                    } else {
+                        Tile.STATE_INACTIVE
+                    }
+                }
             }
         } else {
-            false
+            Tile.STATE_INACTIVE
         }
     }
 
-    private fun updateTileState() {
+    private fun updateTileState(newState: Int) {
         qsTile?.apply {
-            if (secured) {
+            if (newState == Tile.STATE_ACTIVE) {
                 state = Tile.STATE_ACTIVE
                 icon = securedIcon
 
@@ -107,7 +142,6 @@ class MullvadTileService : TileService() {
                     subtitle = resources.getText(R.string.unsecured)
                 }
             }
-
             updateTile()
         }
     }